Adds an experimental feature to hide particular materials.
When invoking vsScene::Draw(), you can now pass in an integer of bit
flags.

Similarly, on materials you can provide a "flags" value.  During render,
we bitwise AND together the flag set on the scene draw with the flags
set on materials;  if the result is non-zero, then the material is
hidden during the draw.

This can be used to quickly cobble together some simple effects, such as
setting some materials as non-shadowcasters by declaring (for example)
that a flag of 1 means "don't cast shadows", and then drawing our shadow
scene with a flag of 1 set (thus hiding all materials with a '1' flags
value set)

This is obviously very early on, but I wanted to experiment with this
sort of functionality, today.