Fix a false positive "memory leak" in the input handling.
We weren't clearing our undo stack after string editing finished until
the next time we edited a string.  This meant that you could wind up
with a "memory leak" being reported for a game, just because that undo
stack hadn't been cleared early enough (it would be cleared later on,
the next time a string was edited)

(There are more mild issues here, such as if that stack has to resize
its storage, etc.  But for a general case, this handles most of the
false positives from here.)