Experimental support for PBO-based asynchronouse loading of vsImage, vsFloatImage
Not fully tested or fully performant yet, but.. it's an interesting
idea, if you want to read back image data that you don't actually need
yet.  This lets you do it asynchronously, and check with the image
object to see when it finishes copying the data to somewhere accessible.

Note that this ONLY currently works for the 'Copy()' interface;
asynch-loaded texture data isn't accessible by the 'GetPixel()'
interfaces.