Add 'IsLoadingContext()' to test whether we're running using the loading context.
This is a little ugly;  honestly, code should be tracking this
themselves, but sometimes you do want to be able to call into a single
function either from a background thread (with the loading context), or
from the main thread.  It's possible that this could be slow on some
processors, although all of my shipping platforms seem to support thread
local storage, so.. probably not *too* bad?