Initial POC for spitting out a backtrace under MINGW.
This simple approach uses dbghlp to print a simple backtrace, up to 25
levels deep.  Note that MinGW does NOT provide symbols;  it's just
printing the PC at each stack level.

This means that I'd need to keep map files around for each build, and
probably write a tool for looking up the stack traces automatically,
so I can decode what's meant by the stack traces.

Which is annoying, but *does* allow me to output backtraces without
actually crashing the program.  And I was wanting to strip debugging
information out of the builds *anyway*.