Address cmake policy CMP0072 (libgl vs libopengl)
Recent CMake builds now support either looking for legacy OpenGL
libraries (libgl.so), or looking for GLVND (libopengl.so).

After minimal testing, the VectorStorm library appears to work just fine
with either library (as one would hope), but cmake really wants us to
specify a preference or it'll spit out warnings.

SO, let's specify that if both are available, then we prefer the modern
GLVND libraries.  We'd be okay with legacy libraries, too, though.

(GLVND is "OpenGL Vendor-Neutral Dispatch")