Removed 'register' storage specifier from vsMatrix4x4::Inverse().
The 'register' specifier was really only ever a compiler hint, and was
aggressively ignored most of the time.  Worse, it's now actually
deprecated as of C++17.

Now, VectorStorm is mostly written against the C++ 98 specifications, so
we don't usually care about changes in more recent C++ standards, but in
this case, clang 7.0+ was complaining about it.  And as the 'register'
specifier wouldn't have actually been making any difference anyhow,
let's just remove it.