Tweaks to justified vsFont rendering.
We now estimate the height of regular capital letters (by grabbing the
height of the 'H' glyph), and we use that when top-justifying or
vertically-center-justifying single-line strings.

This means that when top-justifying a single line of text, the top of
its basic capital letters will now touch y==0 (or whatever position
you've specified for the string), and any portion of a glyph higher than
that will extend further upward, in the same way that descenders will
extend further downward from y==0 when bottom-justifying a string.

This took a silly amount of trial and error to make work!

When justifying multiple lines of text, we use the old behaviour;  use
the whole line height, not just the capital letters.