Fixed out-of-space error when drawing text with a transform.
We always allocated a very small display list (128 bytes) to hold
rendering commands for drawing text, and it turns out that small list
was too small if you specified both a transform for the text and also
wanted a dropshadow.

Doubled the size of that display list to 256 bytes, so it now has space
even in that case.