Fix vsMakeTexturedBox2D_FlipV texture coordinates.
Previously, vsMakeTexturedBox2D_FlipV was making texture coordinates
with 'v' values varying from 0 to -1, instead of 0 to 1.  That worked
fine for normal textures, but if you clamped V values, it totally broke.

Now we correctly set 'v' values from 0 to 1, so clamping works again.

Note that this is potentially a very very mildly breaking change, as an
extra parameter was added to a rarely-used interface.  But in practice,
I only ran into a small number of issues in my very large codebase.