Adds an 'vsAssertF()' function.
Like vsAssert, but accepts printf style string+parameters.

Like vsAssert, the string is not constructed unless the condition check
fails, which means that the expense of the string construction doesn't
matter!  Go nuts, print out any data which would be useful in debugging
the assertion if it should happen to a user!