Adds vsSpline*::Slice() which copies a part of a spline.
This lets you create separate splines which match the position of clips
inside the given spline.

Note that velocity values will not match, as the new spline operates
over a different time interval.  This is the main reason I've added this
utility function, as it's not an immediately intuitive change.

Also marked a bunch of other member functions as const, which hadn't
already been declared that way but should have been.