Adds `DrawMode_MultiplyAbsolute` draw mode.
Our `DrawMode_Multiply` works by multiplying the source and destination
colors together, and then blends the result in by the source alpha
channel.

The new `DrawMode_MultiplyAbsolute` works by multiplying the source
and destination colors together, and just using the resulting value
directly;  no alpha-based modulation.

I feel like this is a bit of a patch, just quickly adding a particular
blending mode that I needed right now, when the proper long-term fix is
to let users directly set blend function parameters in the material,
instead of wrapping these "Draw Modes" around them, even if 99% of the
time the user just wants 'Normal' or 'Add'.

So maybe someday in the future I'll do that?  But for now, this change
gets me to where I need to go today.