Put BIT(x)'s parameter in parentheses.
I can't believe this has been there for so long without either my
noticing or a tool warning me about it!

It's often important to put these macro parameters inside parentheses,
or else bad things can happen!

Here's a real example of a bug caused by lack of parentheses on this
macro.  Imagine you have this code:

```cpp
int value = 5;
int bit = BIT(value-1);
```

In that case, the BIT macro would expand to this: `(1<<value-1)`.  With
'value' being equal to 5, that gives you `1<<5-1` which evaluates to
`(32-1)`, or 31;  you didn't wind up with the fifth bit set, you wound
up with bits one through four set, and NOT the fifth bit!

What we actually wanted was `1<<(5-1)`, which evaluates to `1<<4` which
is 16.  The fourth bit.