Workaround for OSX-Mojave-not-rendering issue.
If we create the OpenGL loading context FIRST rather than SECOND, that
seems to work for getting rendering to work for some reason.

I hate this sort of "jiggle the wires" solution, but right now I just
need something that works, and this is mostly out of our hands;  off
with the SDL folks.