Fixes a bug affecting load binary->save text operations.
If you loaded a binary vsRecord, it was marking all of its labels as
strings.  Which meant that if you saved it in text format, all of those
labels were given quotes around them.

Which would have been fine, except that the parser considers *quoted*
strings to only be strings and NEVER to be labels.  So if you then
reloaded the text version, all of those things which had originally been
labels were now being interpreted as being string tokens in a record
which has no label.

Now we correctly keep label/string type distinct during loading.  Which
fixes my tool for converting between text and binary representations.