Fixes video resolution on first run.
On Windows, on our first-run we created a fullscreen window, and then we
drew only into the bottom left 1024x768 portion of it.  All was fixed on
second run, or if the game changed the video mode itself.

On OSX, on our first run we created a fullscreen window, and then we
drew only into the top left 1024x768 portion of it.

On Linux, on our first run we created a fullscreen window, and then the
window was immediately resized to 1024x768.

Here's what was going on:

With no preferences file, we set a default starting resolution to
1024x768.  But we also decided that by default, we'd create a fullscreen
window, which by definition uses the resolution of the desktop, rather
than the resolution we'd specified.  So here's what was happening
(pseudocode):

We call "SDL_CreateWindow( 1024, 768, FULLSCREENWINDOW );", and SDL
creates us a fullscreen window.  We then used the 1024 and 768 to set
viewports and similar values, so we were only drawing in one corner of
the window.

On Linux, in addition to the above, we would immediately receive a
windows resize event to reset the window size to 1024x768 (presumably
because those were the numbers we'd set in the call to
SDL_CreateWindow()?)

SO.  Fix here is that when we're creating the window, if the window is
"fullscreen window" type, we immediately query back the actual size of
the window so we can create viewports of the correct size.  And then, we
explicitly set the window as being that size;  this seems to fix the
behaviour on Linux where it was immediately resizing back to the
initially specified resolution (which was supposed to be being ignored,
per the docs)