Adds support for asynchronous saving of screenshots.
NOTE:  Asynch screenshot saving occurs by saving the contents of the
offscreen buffer, not the framebuffer itself.  This shouldn't actually
matter, but figured I'd point it out explicitly.

The "Asynch" method means that you're given a vsImage object which is
conducting an asynchronous readback of the screen image.  Check it each
frame until it declares that it has finished reading the image back into
system memory, and then you can do whatever you want with it.

(This is unlike the non-Asynch screenshot functions, which block until
the image has been transferred back to the CPU from the GPU, and so you
can just immediately use the image.  Benefit is that this approach won't
cause frame rate stutters)