Pulls VAO functionality out into vsAttributeBinding.
There's no functional change between before and now;  I'm just in the
process of pulling shader attribute binding out into a new class, in
preparation for having more than one of them.

One of the new OpenGL requirements in 3.x is that you're supposed to
have "Vertex Array Objects" which manage the binding of vertex
attributes. Instead of constantly unbinding and rebinding them manually,
you're supposed to set up objects which hold the complete vertex state,
and just switch from one to the other.  We..  didn't do that.  We just
kept one global VAO and continued to bind and unbind attributes on it as
we rendered.

(In my own defense, at the time that we started doing this, drivers
weren't really optimised for VAO usage, and so what we did was what you
had to do if you wanted to keep performance.  This change is now moving
toward proper handling of VAOs.)

For the moment, though, this is exactly what we were doing before;
we have just one big VAO, and we're setting attributes on it before
every draw.  This is preparatory work before we start having multiple
VAOs, and no longer need to reset all the attributes for each draw.

Other notable thing here:  I've split the vsRendererState into
vsRendererState and vsAttributeState, to pull out the stuff that should
be held within the VAO.  Current thinking is that vsAttributeState is
going to disappear entirely, which will simplify a bunch of interfaces.
But I haven't taken that step yet.  Still working out how I want to
wrangle some of the pieces of this.