Removes old now-unused vsRenderBuffer binding functions.
Also completely removes "RendererState"-style setting of vertex
attribute enabled/disabled state.  That's all now handled *entirely* by
the new vsAttributeBinding class.

This is a terrifying change for me;  it's like a complete reorganisation
of the basic rendering infrastructure of VectorStorm.  Necessary,
though! And MT2 is still running and drawing exactly as it was before.

There is *zero* functional change here.  From OpenGL's perspective, it
wouldn't even be able to tell that we'd modified anything;  Everything
is still working exactly the way it always has, with everything running
through a single "global" VAO object.  Just.. now there's the
architectural possibility to add more, which would mean we wouldn't need
to do as much attribute binding for each draw.