Implemented vsTransform3D multiplication.
It's been there in the interface for years and years, but.. apparently I
never actually implemented it?  Just convert to matrices to do the
multiplication and then pull values back out.

It's not legal to do that for arbitrary matrices, but for ones that were
generated to represent a translation, rotation, and scale (as in this
case), it works just fine.