vsTween now always sets start/end, and works with quaternions.
vsQuaternion now provides a vsInterpolate which does a spherical linear
interpolation.  This allows it to work with vsTween.

vsTween now always sets its 'start' and 'end' values even when not
tweening.  Those values are usually not useful when not tweening, but
could be accessed from the outside via 'GetTarget()', and so I figured I
should probably set them, just so that callers don't need to worry about
the internal implementation of the vsTween.  Just call 'GetTarget()' to
get the desired destination and 'GetValue()' to get the current value.
And both work whether you're tweening or setting explicit values.