Support 3D dropshadow offsets in vsFontRenderer.
This allows you to set a proper 3D offset, so you can give a dropshadow
to 3D text.

Note that x and y components are still expressed in virtual "pixels"
even for 3D text.  The z component of the offset is interpreted as an
absolute depth, regardless of font scale.  (TODO:  Do I want to change
that?)