Failing an assert now closes cleanly.
Also, now compiles, links, and works correctly on modern mingw.

And most importantly, doesn't require distributing libstdc++ DLLs with
the build;  we still link it in statically, if that's what the game
desires.

Also disables printing of debug symbols in MinGW backtraces, since
those are generally unuseful in release builds;  what I really need to
see is the location of the program counter, and previously we were
suppressing that when we thought we had a symbol name.