Fix logging of missing font glyphs.
Previously, we were treating an uninitialised char array as if it was a
"string", and inserting a utf8 character into it, and then trying to
write it out as if it was a regular null-terminated string without ever
having actually written a null there.  This could lead to invalid bytes
in the log!