Remove 'IsDirectory()' check on vsFile constructor.
Profiling shows that PHYSFS_stat is very very slow on Windows, when
operating on the raw file system, and so calling it every time we open
a file was *vastly* increasing file load times on Windows;  like, by a
factor of 50 or more, when loading from an SSD.  (untested for regular
hard drive, but probably similar)

I believe Windows doesn't cache-in-ram the directory structure, and so
the 'stat' call requires going out to disk to get data, unlike OSX or
Linux, which makes this a very costly operation there!

A better long-term option might be moving data into packfiles, which
have a directory structure that can be kept in memory by PhysFS itself,
which would basically make this problem go away, I think?

But for now, just don't do this test any more.  It's been years since
I made that mistake.  And we'll live with the calls to _Stat in
"FileExists" and similar calls.