Proof of concept for a simple file cache.
File performance is still less than awesome on Windows, so this is an
experimental implementation of a file cache;  for particular files
(currently hardcoded as '.win' and '.glsl' files) we remember the
filename and file contents, and if we're asked to open those files again
in the future, instead of actually opening them again, we just provide
the contents we remember from last time we opened the file!

This *ought* to reduce the number of actual files we have to touch,
which should bring up Windows performance by a lot.  (Also, it helps
Linux and OSX, but not by nearly as much since they're already really
fast at these things, so the cache isn't saving as much time there)

Still also worth experimenting with putting data files into packfiles;
then we don't ever have to use the OS functions for opening files at
all, as we're just moving around inside a single large meta-file!

Packfiles are less good for patch-based update systems (Steam, etc), but
if they bring load times down then I'm willing to consider them!