Proof of concept for a "Shader Cache".
We keep track of every shader we load, and if a game asks us to compile
a shader with all the same input data and configuration as one we've
already loaded, we now return a reference to that previously loaded one,
instead of going through the whole compilation process all over again.

Note that this commit is not production-ready;  it leaks all of the
cached shaders rather than closing down properly.  (which, granted,
doesn't actually matter as the whole process will go away afterward
and the driver will clean up for us, but.. it's always nice to clean
up one's own messes rather than leaving them for Mom to do for us)

Initial results:

In MT2, this doesn't result in a lot of saved time.

We're already caching materials, so if two hundred things were using the
"tint" shader, they were already sharing a single shader.  In all, going
from start to sitting on the main map screen we have only 58 shader
loads.  With the cache in place, we have 31 cache hits (where we avoid
compiling a shader) and 27 cache misses (where we do compile a shader).

Those numbers are a lot smaller than I expected!

(side-note:  why am I getting references to "program 654" in
shader recompilation warnings?  We do have a couple other shaders being
compiled not through the material system, for example by the glow
system, but it shouldn't be *that* many more shaders, I'd have thought?)