Switch vsMaterialInternal to use std::atomic<int> and postincrement
Previously it was using the value of a regular int and incrementing
later on, and completely ignoring multithreading issues.

Finally had this one actually cause a problem for me today, as two
materials were assigned the same name, which VectorStorm caught a little
later on.  So.. let's avoid that ever happening again!