Fixes to make shader reloading work again
Now that materials aren't loading shaders directly, we have to
explicitly ask shaders to reload themselves, which we're now doing via
the old historic mechanism that we stopped using when we started
doing full material reloads.

Also, we need to purge the file cache before attempting to re-load those
files we had loaded previously, so we do that too.