Adds vsFontRenderer::SetSizeBias().
This function lets you set a font renderer to use a font size/geometry
for a different size than the size the font is being drawn on screen.

For example, with a size bias of '2', if we ask to draw a font at 12
point, it will take a 24 point version of the font (or the nearest
available), and will rescale it to draw it at a size of 12.

MT2 is using this as part of UI rescaling, to make sure that we keep
crisp text even when the UI scene's coordinates doesn't match the screen
resolution, by forcing us to use higher-quality textures in that case.