Adds 'vsCheck'.
vsCheck is like vsAssert;  it makes a statement that is expected to
evaluate to 'true'.  If not, it prints out a message to the log, just
like with vsAssert, but unlike vsAssert, it doesn't intentionally
crash the program.

Code calling vsCheck is expected to be able to recover from a failure,
where code calling vsAssert will not continue if the condition fails.