Set our process as DPIaware on Windows (even though we're really not)
This makes it so that Windows tells us actual pixels instead of
scaled desktop "points", so even if folks have display scaling turned
on, stuff remains pixel accurate at all times!

SDL2's support for Windows high DPI isn't functional yet, so this is
pretty much the best I can do until that support comes in!  We'll switch
over to that once it's ready to go!