Fix creation of depth and depth/stencil textures to meet OpenGL spec.
We'd inexplicably been declaring them as using a GL_UNSIGNED_BYTE
internal type, which makes zero sense for a texture that's supposed to
contain 24-bit integers and possibly also an 8-bit stencil value.

Changed it to GL_UNSIGNED_INT for depth-only and GL_UNSIGNED_INT_24_8
for depth/stencil.

Hopefully this will fix the problem that one person has reported, on
an Intel chipset!