For the moment, disable the 2D "Is this sprite on-screen" optimisations.
This was written even before VectorStorm was moved into git, and it
totally breaks on ultra-widescreen monitors.  And it probably doesn't
even save us anything.

This could be rewritten to be more accurate and to use the *actual*
display dimensions for the test, instead of this silly "assume it'll be
no wider than 16:9" test.  But for now.. just disable the test entirely!