First working version of vsLocString.
This is my own basic implementation of string formatting, which holds on
to arguments so that a single composed string including all arguments
can be localised multiple times.

This means that we'll be able to have a single composed string and
completely re-localise it if the currently set language changes during
runtime.