Font rendering now *only* works via vsLocString.
We now automatically convert vsStrings to vsLocStrings during the draw
process, so old code should continue to work (but obvs won't support
automatic relocalising)

This commit also restructures how vsLocStrings handle localisation and
literal strings.  We now handle strings for localisation in *exactly*
the same way we handle argument substitution.  If you surround a
substring in {} glyphs, we first try to match an argument against the
substring.  And if that fails, we instead try to match it against the
localisation table.