Adds an interface to return the write directory to the game.
This allows the game to trigger the enging to open up a
Finder/Explorer/etc window showing the write directory.

Also adjusts the Linux version of vsSystem::Launch() to launch xdg-open
in the background, instead of synchronously.  It worked great before for
opening web pages, but works less well when opening other types of
things;  thunar or nautilus or opening text files or whatever.  Now it
happens in the background!