When reporting control strings for binding, consider keyboard layout
Previously, we reported the scancode identifier, which is basically the
US English layout string for the key.  Now, we pass through the OS's
keyboard layout so we show the correct string for the current keyboard
layout.

So if you're using a French AZERTY keyboard, for example, we show that
you'll be using 'Z' to move forward by default, instead of showing 'W'.
It's the same key in the same position on the keyboard, but in the
French layout that key is labelled as 'Z', so let's show it that way
too!