Adds 'globalSeconds' and 'globalMicroseconds' uniform builtins.
'globalTime' is convenient, in that it tells you how many seconds since
the game started running.  However, for long-running games, that value
loses precision over time because it's stored in a float and the value
keeps getting bigger and bigger, which means we start to lose fractional
precision.

'globalSeconds' and 'globalMicroseconds' tell you the number of seconds
since the game started, and how many microseconds into the current
second we're at.  And both values are given as unsigned integers.  This
means that there's no precision loss as numbers get bigger, and we have
enough room to run for a a fair bit more than a century before it rolls
over!

Any shaders using 'globalTime' should probably be migrated to
'globalSeconds' and 'globalMicroseconds', to fix any shader imprecision
errors caused by receiving "chunky time" after the game has been running
for a while.