Undo anisotropy changes.
As it turns out, anisotropy support didn't go into OpenGL until core
context 4.6 (!!), so we need to stick with checking for the extension.

Oops!  Check for support before trying to use it, and use the _EXT
versions of all the interfaces for it.  Also, have our OpenGL startup
trace out whether it's supported, just so it's another useful bit of
data I'll see in logs.