First revision of work toward supporting shader variants.
Idea is that a single "shader" is now a *container* for 'variant'
shaders.  Very much like the vsShaderSuite, but with user-configurable
variants.

In MT2, I'm planning to have variants for:  For/Without lighting,
For/Without rez effect.  Just as an example of what they can be used
for.

In this commit, I've pulled all the real shader functionality into
'vsShaderVariant'. 'vsShader' now creates a vsShaderVariant (which it
calls "current"), and forwards calls through to it.

The next step is going to be setting up a way for the user to specify
variants they're interested in, and for shaders to specify variants they
provide.  And then selecting a variant from the ones available from the
vsShader.  Need to be able to set that variant data both for a full
scene draw, but then also for an individual draw call.  And so these
need to be bits that are kept in common across all the rendered objects
in a game, so the game can say "use the 'drawing into a shadow map'
variant" while it's drawing shadow maps, for example.