Adds the concept of shader "autobits".
With autobits, a game can wire up a uniform value to a shader variant,
so that by setting the uniform value you automatically trigger a bit in
the shader options system.

In MT2, this made it **heaps** easier to just blanket convert all of our
rez effects in different shaders to use shader variants;  no code
changes required.  Just watch for the 'rezEffect' uniform value and
enable the rez effect variant (when available) if it's turned on.