Model instance groups draw per-instance colors using packed colors.
Previously we used full-float color data, which was a little silly.
This reduces per-instance data from 80 bytes to 68 bytes (by converting
the color data from 4 floats to a uint32)

If I'm smart and store the matrix data using position/orientation/scale I
might be able to get our 64 bytes of matrix data down to a position, a
rotation, and a uniform scale.  Which might be 7 or 8 floats, instead of
16.  Which would bring it down to perhaps 32 bytes in total, maybe?  At
a possible cost of more expense in the vertex shader turning the data
back into a matrix.