Workaround for old gcc warning.
GCC was warning that in a HashTable<int>, adding a new element using
operator[] wouldn't initialise that new integer (even though we were
assigning a value to it immediately afterward).

Neither modern GCC or Clang seem to mind any more, but it's cleaner to
do it as an explicit Add with an already-initialised value, instead, so
let's do that.