vsInput now tells SDL where the text field is.
In theory, SDL_SetTextInputRect() is supposed to help inform the OS
exactly where our text box is, so that it can more intelligently
position an IME dialog.

In practice, it doesn't seem to actually do anything.  Still, we might
as well provide the information, right?  This change now requires a game
to tell vsInput where text will appear on the screen, when turning on
string mode.