Const-correct a whole bunch of vsRecord/vsToken functions.
VectorStorm's const-correctness is pretty limited, but I've been going
in and improving it in bits and pieces from time to time, and this is
one of those times.

I've also added a utility 'vsRecord::Label()' function which just
inlines to 'vsRecord::GetLabel()', since it's such a common thing to
want to grab.  Makes calls a little less verbose.