Restructure vsShader to hold a cached copy of its shader text.
This allows a couple nice features.

1.  It allows shaders created in code (i.e.: not loaded from a file) to
    make use of the ShaderVariant system.

2.  It means during a shader reload operation, we only reload each
    shader text once, not once per variant.

Also adjusted vsShader::Reload() to work on system-owned shaders, if
we're using OS-native allocators rather than custom VS allocators.  If
we're using OS-native allocators, we don't use multiple heaps and so we
don't have to worry about whether or not a reloaded system-owned shader
is allocated into the correct heap.