Fixes multisample resolve when stencil/scissor regions are active.
vsRendererState is now a singleton, and we've added the concept of a
"vsRendererStateBlock", which contains state information and can be set
or copied off from the vsRendererState.

This lets a MSAA texture resolve clear out stencil/scissor test status
for the duration of its blit, and then restore to the prior state.

This isn't a full real 'state block' implementation in the traditional
sense (which would include things like partial state data and overlaying
multiple blocks in a queue), but it got me where I needed to go for now,
and is something that I can expand on in the future.