Safety for vsArrayStore and vsVolatileArrayStore RemoveItem()
These array types now remove the item from the array before destroying
it, instead of destroying it before removing it from the array.

This fixes misbehavours where a side-effect of destroying the item
resulted in traversing the array and potentially running into the
partially-destroyed object.

Now the array is always legal to traverse, even while in the middle of
removing an owned object from the array.