When saving single float image to PNG, write to separate channels.
Previously, we wrote a greyscale image, quantised into just 256 values
from the single float value in the image.

Now we instead create a fake RGBA value from the 32 bits of data in the
float.  It doesn't really mean anything and a human can't really
interpret it, but we can at least theoretically load it back in without
losing any data.