Adds vsSquared() inline function.
Just makes it easier in game code to tweak squared values in code,
instead of needing to adjust two copies of the same number when
assigning a squared value.

Just `vsSquared(50.f)` instead of `50.f * 50.f`.  So if you want to
change it to 60.f you only have to change the number once.