Fixes no-bloom rendering on OS X.
On OS X, it's relatively undocumented that you need to bind framebuffer
0 before swapping front and back buffers, or it's swap front and back
buffers of whatever draw framebuffer happened to be bound at the time.

And since Apple declares that no framebuffer other than 0 has front and
back buffers, that turned our swap into a nop, if you happened to have
some other framebuffer set as the draw framebuffer.

In MT2, this would happen if you had 'bloom' turned off.  With 'bloom'
enabled, the last step of our draw was a complicated draw into the
window backbuffer, which set the window as the draw framebuffer.

With bloom *disabled*, the last step of our draw was just a blit into
the window backbuffer, which didn't require actually setting the window
as the draw framebuffer.  So with bloom disabled, we left some other
framebuffer as the current 'draw' framebuffer, and so the 'swap' command
wasn't actually doing anything.

Yay.

It's a fair question as to whether swap was doing anything reasonable in
the other situation, seeing as how I'm just drawing the final image into
it right at the end.  Am I earning anything from drawing into an
offscreen buffer rather than directly into the window's backbuffer?
It's required in order to do multisampling, but other than that?

Hm.

That's a question for another day, though.  This commit fixes no-bloom
rendering on OSX.