Slightly simplify templates in vsCache.
I'm not sure why I was using reinterpret_casts here instead of just
storing the templated object as a direct pointer.  I'm nervous that it
was this way maybe because one of our various platforms didn't like
this for some reason?  But I'm going to give it another try, because
doing it this way makes it so you can use vsResource as part of multiple
inheritance!