Adds 'Reserve()' function to the vsArray classes which were missing it.
"Reserve()" pre-allocates space for the array to hold a particular
number of objects, so it doesn't have to reallocate the array later on
while filling it up.  If you know how many elements are eventually going
to go in the array, Reserve() enough space for them before you start
filling it up!

Also fixes a bug where it was reallocating the array if you specified
an amount to reserve that was higher than the number of items in the
array but *lower* than the amount it had already arranged to be able to
hold.  So for example, if you made an empty array and called Reserve()
twice with the same value, it'd reallocate the array both times!  Now
fixed across all array types.