Experimental implementation of vsFile::Mode_ReadCompressed_Progressive.
Mode_ReadCompressed decompresses the file into RAM immediately.
Mode_ReadCompressed_Progressive instead loads the *compressed* data into
RAM, and then generates and returns decompressed blocks of data from
'ReadBytes()'.  This is *experimental* and probably will break in all
sorts of ways!

However, I've updated the vsSerialiserReadStream to work with it, and
that's the one that's the most critical for MT2 right now.

Long-term we should replace Mode_ReadCompressed with this.  But until
I'm 100% confident in it, I'm going to have it be its own separate mode.