Apply texture clamp settings later on, when deferred-creating textures.
[TODO] is figuring out whether texture clamp settings should be applied
on the texture or on the material.  It's kind of ugly that I have it in
both places right now, with the material (I think?) being universal
across all textures on the material but also with the texture setting
being universal across all materials where the texture is used?  Yuck!