A bunch of extra safety around file writing.
1.  When we try to write data into a file, verify that it got written.

If it fails to write (say, because the drive was out of space), we now
throw an assertion and actually delete the partially-written file, so it
won't cause problems for us in future.

2.  When moving a file (for example, after writing it in a temp
    directory), we now load the temp file into memory, then delete the
    file from disk, then write it in the final position.

Previously we loaded the temp file into memory, write the final file to
disk, and then deleted the temp file.  This meant that we needed twice
the disk space during the move!

(We can't actually do a *real* move, as that's very OS-specific and
isn't exposed through our file access library, so we do a read + delete
+ write instead)