Convert vector, transform, and quaternion classes to return const values for math
This generates compiler errors if you try to (for example) assign a
value to a temporary result.  For example, if you accidentally do:

`-myVector = vsVector3D::One;`

The above code now generates a compiler error instead of how it
previously would silently assign the value to a temporary variable which
would then be destroyed.