Improved vsTween behaviour on smoothed multi-tweens.
When starting a new tween while already inside a tween, smoothed tween objects
would have their velocity start from zero again, in the smoothed change
of value.  To get around this problem, we used to use the old tween
duration, so it was as if we had been tweening toward the newly
specified destination this whole time.  That improved some of the
worst-case situations, but was still pretty terrible, especially when
the new tween was initiated just before the old one completed;  the
tween would just nearly-instantly snap to the new destination.

Instead of that, we now put such tween objects into an "accelerated
tween" mode, where they use slam-in ease-out interpolation, instead of
ease-in ease-out interpolation.  This resolves the problem of a
continually-retweening object appearing to move absurdly slowly!

Of course, if you want *actual* nice behaviour you should totally use
the vsTweenSpline3D, which actually maintains continuity of its velocity
across tweens.  But if you're using another type (quaternions are a
notable one), then the bog standard vsTween now behaves a whole lot
nicer.