Fixes small memory leak after load of some compressed files
When doing a file read using MODE_ReadCompressed_Progressive, we were
leaking the zlib 'zipdata' object when the file was closed.
MODE_ReadCompressed and MODE_WriteCompressed were not affected.

This resulted in a leak of about 7kb of RAM;  not enough to make a
notable difference unless you're loading a *lot* of compressed files
in a progressive mode.  (In MT2, I believe we only load the main save
file in this mode)