Improved processor identification for Apple M1 chipsets.
Since cpuid doesn't work on arm, we instead query sysctl to gather data.
In theory, this seems to be getting us both the CPU's model name ("Apple
M1 Pro", for my laptop), and I've tacked on the number of cores as well.

Hopefully that'll get us a bit more useful information in logfiles!