Proof of concept for setting texture overrides on vsShaderValues objects.
This is kind of halfway toward moving us to a proper shaderblock setup.
I should really just do that someday.  But for now, this is all I really
need for the stuff I'm currently doing, *I think*.

This lets you set an overridden texture on a vsShaderValues object,
which will override what textures were specified on the material being
rendered during the draw.  As with other stuff ont he vsShaderValues,
this works across multiple nested layers of vsShaderValues objects.