[Experimental] moves user files into a separate branch of our file system.
User files now get placed inside a virtual "user" folder.  So if you
want to save out a file, it must be saved into a filename starting with
"user/"

Similarly, if you want to load a file which was saved by the game or
which was otherwise provided by the user, you must load it from a path
starting with "user/"

Only files shipped with the game's own Data directory are mounted into
the root directory, now.  Next step will be to have the game flag
in the log whether any mods are active, and to figure out a nice way for
games to enable/disable mods.