Removed the "allowMods" flag.  Mods are always allowed.
During startup, we now query whether any mods or unpacked data files are
present.  If so, prioritise those files and also print out to the log
that those directories exist.

We no longer mount the working "base" directory as a readable directory
by default.  Games can ask us to make it readable if they need it, but
we no longer offer it by default.